Refactor simulation
At a high level, this commit rearranges the existing simulation components to more closely resemble a conventional simulator, like Gazebo. There is a central simulation node that holds some true state of the vehicle and provides simulated sensor output. The simulation is just the kinematic car model pasted from the robot state node, but includes the basic bounds and map collision checks that were removed from the robot state node.
This prepares the way for tweaks to the simulation (e.g. check for collisions with the footprint, not just the center point, or multivehicle simulation)
The previous robot state-hosted simulation included some sly TF backfilling. Whenever it detected that noone was providing map->odom, it would publish the ground truth data. This isn't how a real robot behaves, so it can be misleading. This refactor replaces the previous "provide if no one is providing" with an explicit launch flag to enable a `fake_localization` node. This node consumes the simulator's ground truth odometry data and offers the conventional map->odom transform.
The launch files have been updated to support the new structure.
The vesc provided odometry node is dropped, since it wasn't clear if anyone consumed its data (which seems to be based on a less refined kinematic model)
To align with typical ROS navigation behavior, the "2D Pose Estimate" `/initialpose` Rviz tool is reserved for initializing a localization provider. I added a special node to republish the "/clicked_point" output as the new command to reposition the robot in the simulation